authorization

+++ backend/routes/posts.js
@@ -35,7 +35,8 @@ router.post('',
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId,
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -71,9 +72,14 @@ router.get('', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      res.status(200).json({ message: 'Post deleted successfully'});
+      if (result.n > 0) {
+        res.status(200).json({message: 'Post deleted successfully'});
+      }
+      else {
+        res.status(401).json({message: 'Not authorized'});
+      }
     });
 });
 
@@ -90,12 +96,17 @@ router.put('/:id',
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId,
   });
-  Post.updateOne({_id: req.params.id}, post)
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
     .then(result => {
-      console.log(result);
-      res.status(200).json({message: 'Post updated successfully'});
+      if (result.nModified > 0) {
+        res.status(200).json({message: 'Post updated successfully'});
+      }
+      else {
+        res.status(401).json({message: 'Not authorized'});
+      }
     })
 });
